pyproject.toml
[project]
name = "alert-app"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "absl-py==2.1.0",
    "accelerate==1.12.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.3",
    "aiosignal==1.4.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "async-timeout==5.0.1",
    "asyncclick==8.3.0.3",
    "attrs==24.2.0",
    "build==1.4.0",
    "certifi==2024.8.30",
    "cffi==2.0.0",
    "charset-normalizer==3.4.0",
    "click==8.3.1",
    "contourpy==1.3.1",
    "cryptography==46.0.5",
    "cvzone==1.6.1",
    "cycler==0.12.1",
    "exceptiongroup==1.3.1",
    "filelock==3.16.1",
    "flatbuffers==24.3.25",
    "fonttools==4.55.3",
    "frozenlist==1.8.0",
    "fsspec==2024.10.0",
    "h11==0.16.0",
    "hf-xet==1.3.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "huggingface_hub==1.4.1",
    "idna==3.10",
    "importlib_metadata==8.7.1",
    "jax==0.4.37",
    "jaxlib==0.4.36",
    "Jinja2==3.1.4",
    "kiwisolver==1.4.7",
    "lap==0.5.12",
    "markdown-it-py==4.0.0",
    "MarkupSafe==3.0.2",
    "matplotlib==3.9.3",
    "mdurl==0.1.2",
    "mediapipe==0.10.18",
    "ml_dtypes==0.5.0",
    "mpmath==1.3.0",
    "multidict==6.7.1",
    "networkx==3.4.2",
    "numpy==1.26.4",
    "nvidia-cublas-cu12==12.4.5.8",
    "nvidia-cuda-cupti-cu12==12.4.127",
    "nvidia-cuda-nvrtc-cu12==12.4.127",
    "nvidia-cuda-runtime-cu12==12.4.127",
    "nvidia-cudnn-cu12==9.1.0.70",
    "nvidia-cufft-cu12==11.2.1.3",
    "nvidia-curand-cu12==10.3.5.147",
    "nvidia-cusolver-cu12==11.6.1.9",
    "nvidia-cusparse-cu12==12.3.1.170",
    "nvidia-nccl-cu12==2.21.5",
    "nvidia-nvjitlink-cu12==12.4.127",
    "nvidia-nvtx-cu12==12.4.127",
    "opencv-contrib-python==4.10.0.84",
    "opencv-python==4.10.0.84",
    "opt_einsum==3.4.0",
    "packaging==24.2",
    "paho-mqtt==2.1.0",
    "pandas==2.2.3",
    "pillow==11.0.0",
    "pip-tools==7.5.3",
    "propcache==0.4.1",
    "protobuf==4.25.5",
    "psutil==6.1.0",
    "psycopg2-binary==2.9.11",
    "py-cpuinfo==9.0.0",
    "pycparser==2.22",
    "pycryptodome==3.23.0",
    "pydantic==1.10.26",
    "pydantic_core==2.41.5",
    "Pygments==2.19.2",
    "pyparsing==3.2.0",
    "pyproject_hooks==1.2.0",
    "pytapo==3.4.11",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-kasa==0.5.3",
    "pytz==2024.2",
    "PyYAML==6.0.2",
    "redis==7.2.1",
    "regex==2024.11.6",
    "requests==2.32.3",
    "rich==14.3.3",
    "rtp==0.0.4",
    "safetensors==0.4.5",
    "scipy==1.14.1",
    "seaborn==0.13.2",
    "sentencepiece==0.2.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sounddevice==0.5.1",
    "sqlparse==0.5.5",
    "sympy==1.13.1",
    "tabulate==0.10.0",
    "tokenizers==0.22.2",
    "tomli==2.4.0",
    "torch==2.5.1",
    "torchvision==0.20.1",
    "tqdm==4.67.1",
    "transformers==5.2.0",
    "triton==3.1.0",
    "typer==0.24.1",
    "typer-slim==0.24.0",
    "types-psycopg2==2.9.21.20260223",
    "typing-inspection==0.4.2",
    "typing_extensions==4.15.0",
    "tzdata==2024.2",
    "ultralytics==8.3.49",
    "ultralytics-thop==2.0.13",
    "urllib3==2.2.3",
    "websocket-client==1.9.0",
    "yarl==1.23.0",
    "yoyo-migrations==9.0.0",
    "zipp==3.23.0",
]

[tool.uv]
extra-index-url = ["https://download.pytorch.org/whl/cu124"]

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
